screwdriver act uses tool proc!

+++ code/game/machinery/lightswitch.dm
@@ -65,6 +65,8 @@
 
 /obj/machinery/light_switch/screwdriver_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
 	var/obj/item/frame/light_switch/frame = new /obj/item/frame/light_switch(user.loc, 1)
 	transfer_fingerprints_to(frame)
 	qdel(src)
