+++ code/modules/power/lighting.dm
@@ -65,32 +65,34 @@
 
 /obj/machinery/light_construct/screwdriver_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
-	if (stage == LIGHT_STAGE_WIRED)
-		stage = LIGHT_STAGE_COMPLETE
-		update_icon()
-		user.visible_message(
-			SPAN_NOTICE("\The [user] closes \the [src]'s casing."),
-			SPAN_NOTICE("You close \the [src]'s casing."),
-			SPAN_ITALIC("You hear screws being tightened.")
-		)
-		playsound(loc, 'sound/items/Screwdriver.ogg', 50, TRUE)
-		var/obj/machinery/light/newlight = new fixture_type(loc, src)
-		newlight.set_dir(dir)
-		transfer_fingerprints_to(newlight)
-		qdel(src)
+	if(stage != LIGHT_STAGE_WIRED)
+		return
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
+	stage = LIGHT_STAGE_COMPLETE
+	update_icon()
+	user.visible_message(
+		SPAN_NOTICE("[user] closes [src]'s casing."),
+		SPAN_NOTICE("You close [src]'s casing."),
+		SPAN_ITALIC("You hear screws being tightened.")
+	)
+	var/obj/machinery/light/newlight = new fixture_type(loc, src)
+	newlight.set_dir(dir)
+	transfer_fingerprints_to(newlight)
+	qdel(src)
 
 /obj/machinery/light_construct/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
 	switch(stage)
 		if (LIGHT_STAGE_EMPTY)
 			playsound(loc, 'sound/items/Ratchet.ogg', 50, TRUE)
-			to_chat(user, SPAN_NOTICE("You begin deconstructing \the [src]."))
+			to_chat(user, SPAN_NOTICE("You begin deconstructing [src]."))
 			if (!user.do_skilled((tool.toolspeed * 3) SECONDS, SKILL_CONSTRUCTION, src, do_flags = DO_REPAIR_CONSTRUCT))
 				return
 			new /obj/item/stack/material/steel( get_turf(loc), sheets_refunded )
 			user.visible_message(
-				SPAN_NOTICE("\The [user] deconstructs \the [src]."),
-				SPAN_NOTICE("You deconstruct \the [src]!"),
+				SPAN_NOTICE("[user] deconstructs [src]."),
+				SPAN_NOTICE("You deconstruct [src]!"),
 				SPAN_ITALIC("You hear ratcheting and metal scraping.")
 			)
 			playsound(loc, 'sound/items/Deconstruct.ogg', 75, TRUE)
@@ -114,8 +116,8 @@
 	update_icon()
 	new /obj/item/stack/cable_coil(get_turf(loc), 1, "red")
 	user.visible_message(
-		SPAN_NOTICE("\The [user] cuts the wires from \the [src]."),
-		SPAN_NOTICE("You cut \the [src]'s wires and remove them from the frame'."),
+		SPAN_NOTICE("[user] cuts the wires from [src]."),
+		SPAN_NOTICE("You cut [src]'s wires and remove them from the frame'."),
 		SPAN_ITALIC("You hear snipping and cables being cut.")
 	)
 	playsound(loc, 'sound/items/Wirecutter.ogg', 50, TRUE)
@@ -130,8 +132,8 @@
 			stage = LIGHT_STAGE_WIRED
 			update_icon()
 			user.visible_message(
-				SPAN_NOTICE("\The [user] adds wires to \the [src]."),
-				SPAN_NOTICE("You add wires to \the [src].")
+				SPAN_NOTICE("[user] adds wires to [src]."),
+				SPAN_NOTICE("You add wires to [src].")
 			)
 			playsound(loc, 'sound/items/Deconstruct.ogg', 50, TRUE)
 		return TRUE
@@ -339,13 +341,13 @@
 		return
 	var/status = get_status()
 	if(status == LIGHT_EMPTY || status == LIGHT_BROKEN)
-		to_chat(user, SPAN_WARNING("You can't do anything with \the [src]."))
+		to_chat(user, SPAN_WARNING("You can't do anything with [src]."))
 		return
 	if(!(status == LIGHT_OK || status == LIGHT_BURNED))
 		return
 	user.visible_message(
-		SPAN_WARNING("\The [user] smashes \the [src]!"),
-		SPAN_WARNING("You smash \the [src]!")
+		SPAN_WARNING("[user] smashes [src]!"),
+		SPAN_WARNING("You smash [src]!")
 	)
 	attack_animation(user)
 	broken()
@@ -425,8 +427,8 @@
 		user.do_attack_animation(src)
 		if(prob(1 + weapon.force * 5))
 			user.visible_message(
-				SPAN_WARNING("\The [user] smashes \the [src]!"),
-				SPAN_WARNING("You smash \the [src]!"),
+				SPAN_WARNING("[user] smashes [src]!"),
+				SPAN_WARNING("You smash [src]!"),
 				SPAN_WARNING("You hear a small glass object shatter.")
 			)
 			if(on && (weapon.obj_flags & OBJ_FLAG_CONDUCTIBLE))
@@ -435,8 +437,8 @@
 			broken()
 		else
 			user.visible_message(
-				SPAN_WARNING("\The [user] hits \the [src]!"),
-				SPAN_WARNING("You hit \the [src]!"),
+				SPAN_WARNING("[user] hits [src]!"),
+				SPAN_WARNING("You hit [src]!"),
 				SPAN_WARNING("You hear glass cracking.")
 			)
 			playsound(loc, "glasscrack", 40, TRUE)
@@ -445,13 +447,14 @@
 	return ..()
 
 /obj/machinery/light/screwdriver_act(mob/living/user, obj/item/tool)
-	. = ..()
+	. = ITEM_INTERACT_SUCCESS
 	if(lightbulb)
 		return
-	playsound(loc, 'sound/items/Screwdriver.ogg', 50, TRUE)
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
 	user.visible_message(
-		SPAN_NOTICE("\The [user] opens \the [src]'s casing."),
-		SPAN_NOTICE("You open up \the [src]'s casing."),
+		SPAN_NOTICE("[user] opens [src]'s casing."),
+		SPAN_NOTICE("You open up [src]'s casing."),
 		SPAN_ITALIC("You hear screws being loosened.")
 	)
 	var/obj/machinery/light_construct/C = new construct_type(loc, dir, src)
@@ -471,8 +474,8 @@
 		if(!user.unEquip(W, src))
 			return TRUE
 		user.visible_message(
-			SPAN_NOTICE("\The [user] inserts \a [W] into \the [src]."),
-			SPAN_NOTICE("You insert \the [W] into \the [src]."),
+			SPAN_NOTICE("[user] inserts \a [W] into [src]."),
+			SPAN_NOTICE("You insert [W] into [src]."),
 			SPAN_ITALIC("You hear something being screwed in.")
 		)
 		insert_bulb(W)
@@ -480,8 +483,8 @@
 
 	if (!lightbulb)
 		user.visible_message(
-			SPAN_WARNING("\The [user] shoves \a [W] into \the [src]!"),
-			SPAN_DANGER("You stick \the [W] into \the [src]!")
+			SPAN_WARNING("[user] shoves \a [W] into [src]!"),
+			SPAN_DANGER("You stick [W] into [src]!")
 		)
 		if(powered() && (W.obj_flags & OBJ_FLAG_CONDUCTIBLE))
 			var/datum/effect/spark_spread/s = new /datum/effect/spark_spread
@@ -518,7 +521,7 @@
 // ai attack - make lights flicker, because why not
 
 /obj/machinery/light/attack_ai(mob/user)
-	to_chat(user, SPAN_NOTICE("You cause \the [src] to flick on and off."))
+	to_chat(user, SPAN_NOTICE("You cause [src] to flick on and off."))
 	flicker(1)
 
 // attack with hand - remove tube/bulb
@@ -532,8 +535,8 @@
 		var/mob/living/carbon/human/H = user
 		if(H.species.can_shred(H))
 			H.visible_message(
-				SPAN_WARNING("\The [user] shreds \the [src]!"),
-				SPAN_WARNING("You shred \the [src]!"),
+				SPAN_WARNING("[user] shreds [src]!"),
+				SPAN_WARNING("You shred [src]!"),
 				SPAN_WARNING("You hear glass shattering!"))
 			broken()
 			return TRUE
@@ -561,7 +564,7 @@
 			if(hand && hand.is_usable() && !hand.can_feel_pain())
 				user.apply_damage(3, DAMAGE_BURN, user.hand ? BP_L_HAND : BP_R_HAND, used_weapon = src)
 				user.visible_message(
-					SPAN_WARNING("\The [user]'s [hand] burns and sizzles as \he touches the hot [get_fitting_name()]."),
+					SPAN_WARNING("[user]'s [hand] burns and sizzles as \he touches the hot [get_fitting_name()]."),
 					SPAN_WARNING("Your [hand.name] burns and sizzles as you remove the hot [get_fitting_name()].")
 				)
 		else
@@ -684,7 +687,7 @@
 /obj/item/light/examine(mob/user)
 	. = ..()
 	if (reagents?.total_volume && Adjacent(user))
-		to_chat(user, SPAN_WARNING("There's some sort of fluid inside \the [src]."))
+		to_chat(user, SPAN_WARNING("There's some sort of fluid inside [src]."))
 
 /obj/item/light/get_color()
 	return b_colour
@@ -834,12 +837,12 @@
 			if (!reagents)
 				create_reagents(5)
 				S.reagents.trans_to_obj(src, 5)
-				to_chat(user, SPAN_WARNING("You inject the solution into \the [src]."))
+				to_chat(user, SPAN_WARNING("You inject the solution into [src]."))
 				if (reagents.get_reagent_amount(/datum/reagent/toxin/phoron) >= LIGHT_PHORON_EXPLODE_THRESHOLD)
 					log_and_message_admins("injected a light with phoron, rigging it to explode.", user)
 				return
 			else
-				to_chat(user, SPAN_WARNING("\The [src] is already filled with fluid!"))
+				to_chat(user, SPAN_WARNING("[src] is already filled with fluid!"))
 	. = ..()
 
 // shatter light, unless it was an attempt to put it in a light socket
@@ -867,7 +870,7 @@
 /obj/item/light/proc/shatter()
 	if(status == LIGHT_OK || status == LIGHT_BURNED)
 		visible_message(
-			SPAN_WARNING("\The [src] shatters!"),
+			SPAN_WARNING("[src] shatters!"),
 			SPAN_WARNING("You hear a small glass object shatter.")
 		)
 		set_status(LIGHT_BROKEN)
@@ -879,7 +882,7 @@
 	if(reagents)
 		if (reagents.get_reagent_amount(/datum/reagent/toxin/phoron) >= LIGHT_PHORON_EXPLODE_THRESHOLD)
 			visible_message(
-				SPAN_DANGER("\The [src] flares brilliantly!"),
+				SPAN_DANGER("[src] flares brilliantly!"),
 				SPAN_DANGER("You hear a loud crack!")
 			)
 			log_and_message_admins("Rigged light explosion, last touched by [fingerprintslast]")
@@ -887,7 +890,7 @@
 			set_status(LIGHT_BROKEN)
 			addtimer(CALLBACK(GLOBAL_PROC, GLOBAL_PROC_REF(explosion), T, 3, EX_ACT_LIGHT), 0.5 SECONDS)
 		else
-			visible_message(SPAN_WARNING("\The [src] short-circuits as something burns out its filament!"))
+			visible_message(SPAN_WARNING("[src] short-circuits as something burns out its filament!"))
 			set_status(LIGHT_BURNED)
 			if (sound_on)
 				playsound(src, sound_on, 100)
